+++ Control_C_ADS_Touch_3_5/platylogo.h
